refactor: Split advanced_objects.rs parser into focused modules
Split the 522-line advanced_objects.rs file into object-specific modules
to improve organization, maintainability, and parallel development.

Changes:
- Created crates/parser/src/parser/advanced_objects/ directory
- Split into 6 focused modules:
  - sequence.rs (~190 lines) - SEQUENCE parsers
  - type_objects.rs (~70 lines) - TYPE parsers
  - collation.rs (~130 lines) - COLLATION + CHARACTER SET parsers
  - translation.rs (~60 lines) - TRANSLATION parsers
  - assertion.rs (~50 lines) - ASSERTION parsers
  - mod.rs - Re-exports all functions for backward compatibility
- Removed original advanced_objects.rs file

Benefits:
- Focused modules: Each file handles one SQL object type
- Better discoverability: Clear which file contains which parser
- Parallel development: Different developers can work on different parsers
- Easier testing: Can add parser tests alongside implementation
- Maintainability: Smaller files are easier to understand and modify

Impact:
- Zero runtime performance impact
- No breaking changes (all functions re-exported from mod.rs)
- All 716 parser tests pass
- No API changes

Closes #788
